Fix an issue where a canceled request causes the test to fail
There is a page that leverages Leaflet component. When rendering the map layer, Leaflet cancels the request for a specific resource and requests it again. The cancellation causes the Loki test to fail. The error message is '1 request failed to load; https://a.tile.openstreetmap.org/0/0/0.png'.

It seems we should consider the canceled request as a normal ended request rather than a failed request.